refactor: remove Settings class and factory, integrate settings directly in WishManagerFactory
