+++ wish-models/src/wish_models/test_factories/__init__.py
@@ -2,6 +2,7 @@ from .command_result_factory import CommandResultDoingFactory, CommandResultSucc
 from .executable_collection_factory import ExecutableCollectionFactory
 from .executable_info_factory import ExecutableInfoFactory
 from .log_files_factory import LogFilesFactory
+from .settings_factory import SettingsFactory
 from .system_info_factory import SystemInfoFactory
 from .utc_datetime_factory import UtcDatetimeFactory
 from .wish_factory import WishDoingFactory, WishDoneFactory
@@ -10,6 +11,7 @@ __all__ = [
     "CommandResultDoingFactory",
     "CommandResultSuccessFactory",
     "LogFilesFactory",
+    "SettingsFactory",
     "UtcDatetimeFactory",
     "WishDoingFactory",
     "WishDoneFactory",
