+++ wish-sh/tests/test_wish_paths.py
@@ -1,7 +1,8 @@
 from pathlib import Path
 from unittest.mock import mock_open, patch
 
-from wish_sh.test_factories.settings_factory import SettingsFactory
+from wish_models.test_factories import SettingsFactory
+
 from wish_sh.wish_paths import WishPaths
